+++ src/egon/data/airflow/dags/pipeline.py
@@ -27,8 +27,14 @@ from egon.data.datasets.electricity_demand_timeseries import (
 from egon.data.datasets.electricity_demand_timeseries.cts_buildings import (
     CtsDemandBuildings,
 )
+from egon.data.datasets.emobility.heavy_duty_transport import (
+    HeavyDutyTransport,
+)
 from egon.data.datasets.emobility.motorized_individual_travel import (
     MotorizedIndividualTravel,
+)
+from egon.data.datasets.emobility.motorized_individual_travel_charging_infrastructure import (  # noqa: E501
+    MITChargingInfrastructure,
 )
 from egon.data.datasets.era5 import WeatherData
 from egon.data.datasets.etrago_setup import EtragoSetup
@@ -562,6 +568,15 @@ with airflow.DAG(
             gas_production_insert_data,
             insert_data_ch4_storages,
         ]
+    )
+
+    mit_charging_infrastructure = MITChargingInfrastructure(
+        dependencies=[mv_grid_districts, hh_demand_buildings_setup]
+    )
+
+    # eMobility: heavy duty transport
+    heavy_duty_transport = HeavyDutyTransport(
+        dependencies=[vg250, setup_etrago, create_gas_polygons_egon2035]
     )
 
     cts_demand_buildings = CtsDemandBuildings(